+++ library/src/engine/index.ts
@@ -1,4 +1,5 @@
 import { Computed } from "../plugins/official/core/attributes/computed";
+import { Modifiers } from "../plugins/official/core/attributes/modifiers";
 import { Signals } from "../plugins/official/core/attributes/signals";
 import { Star } from "../plugins/official/core/attributes/star";
 import { SignalValueMacro } from "../plugins/official/core/macros/signals";
@@ -8,6 +9,7 @@ const ds = new Engine();
 ds.load(
     Star,
     SignalValueMacro,
+    Modifiers,
     Signals,
     Computed,
 );