+++ library/src/plugins/official/core/attributes/signals.ts
@@ -4,6 +4,7 @@ import {
     PluginType,
     Requirement,
 } from "../../../../engine/types";
+import { modifiers } from "../../../../utils/modifiers";
 import { jsStrToObject } from "../../../../utils/text";
 
 export const Signals: AttributePlugin = {
@@ -13,12 +14,15 @@ export const Signals: AttributePlugin = {
     removeOnLoad: true,
     onLoad: (ctx) => {
         const { key, genRX, signals } = ctx;
+        const mods = modifiers(ctx);
         if (key != "") {
             signals.setValue(key, genRX()());
         } else {
             const obj = jsStrToObject(ctx.value);
             ctx.value = JSON.stringify(obj);
-            signals.merge(genRX()<NestedValues>());
+            const ifMissing = mods?.onlyifmissing ?? false;
+
+            signals.merge(genRX()<NestedValues>(), ifMissing);
         }
     },
 };