+++ sdk/dotnet/src/Consts.fs
@@ -38,8 +38,8 @@ type EventType =
 module Consts =
     let [<Literal>] DatastarKey               = "datastar"
     let [<Literal>] Version                   = "0.21.0-beta2"
-    let [<Literal>] VersionClientByteSize     = 33647
-    let [<Literal>] VersionClientByteSizeGzip = 12356
+    let [<Literal>] VersionClientByteSize     = 32643
+    let [<Literal>] VersionClientByteSizeGzip = 11947
 
     /// Default: TimeSpan.FromMilliseconds 300
     let DefaultSettleDuration = TimeSpan.FromMilliseconds 300