+++ sdk/go/consts.go
@@ -7,8 +7,8 @@ import "time"
 const (
     DatastarKey = "datastar"
     Version                   = "0.21.0-beta2"
-    VersionClientByteSize     = 33647
-    VersionClientByteSizeGzip = 12356
+    VersionClientByteSize     = 32643
+    VersionClientByteSizeGzip = 11947
 
     //region Default durations